// src/server/eventsub-types.ts
type EventSubTypeEntity<V extends number, C extends Record<string, Condition<any, any>>, P = unknown> = {
  conditions: C;
  version: V;
  payload: P;
};

export type EventSubType = {
  'channel.follow': EventSubTypeEntity<2, {
    broadcaster_user_id: Condition;
    moderator_user_id: Condition;
  }, {
    user_id: string;
    user_login: string;
    user_name: string;
    broadcaster_user_id: string;
    broadcaster_user_login: string;
    broadcaster_user_name: string;
    followed_at: string;
  }>

  'channel.update': EventSubTypeEntity<2, {
    broadcaster_user_id: Condition;
  }>;

  'channel.chat.message': EventSubTypeEntity<1, {
    broadcaster_user_id: Condition;
    user_id: Condition;
  }, {
    broadcaster_user_id: string;
    broadcaster_user_name: string;
    broadcaster_user_login: string;
    chatter_user_id: string;
    chatter_user_name: string;
    chatter_user_login: string;
    message_id: string;
    message: {
      text: string;
      fragments: Array<{
        type: 'text' | 'cheermote' | 'emote' | 'mention';
        text: string;
        cheermote?: null | {
          prefix: string;
          bits: number;
          tier: number;
        };
        emote?: null | {
          id: string;
          emote_set_id: string;
          owner_id: string;
          format: Array<'static' | 'animated'>;
        };
        mention?: null | {
          user_id: string;
          user_name: string;
          user_login: string;
        };
      }>;
    };
    message_type: 'text' | 'channel_points_highlighted' | 'channel_points_sub_only' | 'user_intro' | 'power_ups_message_effect' | 'power_ups_gigantified_emote';
    badges: Array<{
      set_id: string;
      id: string;
      info: string;
    }>;
    cheer?: null | {
      bits: number;
    };
    color: string;
    reply?: null | {
      parent_message_id: string; //	An ID that uniquely identifies the parent message that this message is replying to.
      parent_message_body: string; //	The message body of the parent message.
      parent_user_id: string; //	User ID of the sender of the parent message.
      parent_user_name: string; //	User name of the sender of the parent message.
      parent_user_login: string; //	User login of the sender of the parent message.
      thread_message_id: string; //	An ID that identifies the parent message of the reply thread.
      thread_user_id: string; //	User ID of the sender of the thread’s parent message.
      thread_user_name: string; //	User name of the sender of the thread’s parent message.
      thread_user_login: string; //	User login of the sender of the thread’s parent message.
    };
    channel_points_custom_reward_id: null | string;
    source_broadcaster_user_id: null | string;
    source_broadcaster_user_name: null | string;
    source_broadcaster_user_login: null | string;
    source_message_id: null | string;
    source_badges?: null | {
      set_id: string;
      id: string;
      info: string;
    };
    is_source_only?: null | boolean;
  }>;

  'channel.chat.message_delete': EventSubTypeEntity<1, {
    broadcaster_user_id: Condition;
    user_id: Condition;
  }>;

  'channel.chat.notification': EventSubTypeEntity<1, {
    broadcaster_user_id: Condition;
    user_id: Condition;
  }>;

  'channel.subscribe': EventSubTypeEntity<1, {
    broadcaster_user_id: Condition;
  }, {
    user_id: string;
    user_login: string;
    user_name: string;
    broadcaster_user_id: string;
    broadcaster_user_login: string;
    broadcaster_user_name: string;
    tier: string;
    is_gift: boolean;
  }>;

  'channel.raid': EventSubTypeEntity<1, {
    from_broadcaster_user_id: Condition<string, false>;
    to_broadcaster_user_id: Condition<string, false>;
  }, {
    from_broadcaster_user_id: string;
    from_broadcaster_user_login: string;
    from_broadcaster_user_name: string;
    to_broadcaster_user_id: string;
    to_broadcaster_user_login: string;
    to_broadcaster_user_name: string;
    viewers: number;
  }>;
};